EVERYTHINGIS64BITSOTHEREISNOMOREISSUESWITHNEGITIVENUMBERS
So my friend broke the game by getting a freaking 11 billion times
multipler for the kek, then when he clicked it it flipped the smiles to
negitive..... i had a 32 bit int and it died, what am i, gangnem style?!
ok everything is long and fixed. also a minor error was fixed too.